+++ tests/Jroc.Test262.Tests/language/identifiers/Test262BatchPort20260919Round4ExecutionTests.cs
@@ -1,0 +1,261 @@
+using Jroc.Test262.Tests.language;
+
+namespace Jroc.Test262.Tests.language.identifiers;
+
+public class Test262BatchPort20260919Round4ExecutionTests : DiskExecutionTestsBase
+{
+    public Test262BatchPort20260919Round4ExecutionTests() : base("language.identifiers") { }
+
+    [Fact(DisplayName = "start-unicode-14.0.0-class-escaped")]
+    public Task start_unicode_14_0_0_class_escaped()
+        => ExecutionTest("start-unicode-14.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-14.0.0-class")]
+    public Task start_unicode_14_0_0_class()
+        => ExecutionTest("start-unicode-14.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-14.0.0-escaped")]
+    public Task start_unicode_14_0_0_escaped()
+        => ExecutionTest("start-unicode-14.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-14.0.0")]
+    public Task start_unicode_14_0_0()
+        => ExecutionTest("start-unicode-14.0.0");
+
+    [Fact(DisplayName = "start-unicode-15.0.0-class-escaped")]
+    public Task start_unicode_15_0_0_class_escaped()
+        => ExecutionTest("start-unicode-15.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-15.0.0-class")]
+    public Task start_unicode_15_0_0_class()
+        => ExecutionTest("start-unicode-15.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-15.0.0-escaped")]
+    public Task start_unicode_15_0_0_escaped()
+        => ExecutionTest("start-unicode-15.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-15.0.0")]
+    public Task start_unicode_15_0_0()
+        => ExecutionTest("start-unicode-15.0.0");
+
+    [Fact(DisplayName = "start-unicode-15.1.0-class-escaped")]
+    public Task start_unicode_15_1_0_class_escaped()
+        => ExecutionTest("start-unicode-15.1.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-15.1.0-class")]
+    public Task start_unicode_15_1_0_class()
+        => ExecutionTest("start-unicode-15.1.0-class");
+
+    [Fact(DisplayName = "start-unicode-15.1.0-escaped")]
+    public Task start_unicode_15_1_0_escaped()
+        => ExecutionTest("start-unicode-15.1.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-15.1.0")]
+    public Task start_unicode_15_1_0()
+        => ExecutionTest("start-unicode-15.1.0");
+
+    [Fact(DisplayName = "start-unicode-16.0.0-class-escaped")]
+    public Task start_unicode_16_0_0_class_escaped()
+        => ExecutionTest("start-unicode-16.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-16.0.0-class")]
+    public Task start_unicode_16_0_0_class()
+        => ExecutionTest("start-unicode-16.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-16.0.0-escaped")]
+    public Task start_unicode_16_0_0_escaped()
+        => ExecutionTest("start-unicode-16.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-16.0.0")]
+    public Task start_unicode_16_0_0()
+        => ExecutionTest("start-unicode-16.0.0");
+
+    [Fact(DisplayName = "start-unicode-17.0.0-class-escaped")]
+    public Task start_unicode_17_0_0_class_escaped()
+        => ExecutionTest("start-unicode-17.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-17.0.0-class")]
+    public Task start_unicode_17_0_0_class()
+        => ExecutionTest("start-unicode-17.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-17.0.0-escaped")]
+    public Task start_unicode_17_0_0_escaped()
+        => ExecutionTest("start-unicode-17.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-17.0.0")]
+    public Task start_unicode_17_0_0()
+        => ExecutionTest("start-unicode-17.0.0");
+
+    [Fact(DisplayName = "start-unicode-5.2.0-class-escaped")]
+    public Task start_unicode_5_2_0_class_escaped()
+        => ExecutionTest("start-unicode-5.2.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-5.2.0-class")]
+    public Task start_unicode_5_2_0_class()
+        => ExecutionTest("start-unicode-5.2.0-class");
+
+    [Fact(DisplayName = "start-unicode-5.2.0-escaped")]
+    public Task start_unicode_5_2_0_escaped()
+        => ExecutionTest("start-unicode-5.2.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-5.2.0")]
+    public Task start_unicode_5_2_0()
+        => ExecutionTest("start-unicode-5.2.0");
+
+    [Fact(DisplayName = "start-unicode-6.0.0-class-escaped")]
+    public Task start_unicode_6_0_0_class_escaped()
+        => ExecutionTest("start-unicode-6.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-6.0.0-class")]
+    public Task start_unicode_6_0_0_class()
+        => ExecutionTest("start-unicode-6.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-6.0.0-escaped")]
+    public Task start_unicode_6_0_0_escaped()
+        => ExecutionTest("start-unicode-6.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-6.0.0")]
+    public Task start_unicode_6_0_0()
+        => ExecutionTest("start-unicode-6.0.0");
+
+    [Fact(DisplayName = "start-unicode-6.1.0-class-escaped")]
+    public Task start_unicode_6_1_0_class_escaped()
+        => ExecutionTest("start-unicode-6.1.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-6.1.0-class")]
+    public Task start_unicode_6_1_0_class()
+        => ExecutionTest("start-unicode-6.1.0-class");
+
+    [Fact(DisplayName = "start-unicode-6.1.0-escaped")]
+    public Task start_unicode_6_1_0_escaped()
+        => ExecutionTest("start-unicode-6.1.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-6.1.0")]
+    public Task start_unicode_6_1_0()
+        => ExecutionTest("start-unicode-6.1.0");
+
+    [Fact(DisplayName = "start-unicode-7.0.0-class-escaped")]
+    public Task start_unicode_7_0_0_class_escaped()
+        => ExecutionTest("start-unicode-7.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-7.0.0-class")]
+    public Task start_unicode_7_0_0_class()
+        => ExecutionTest("start-unicode-7.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-7.0.0-escaped")]
+    public Task start_unicode_7_0_0_escaped()
+        => ExecutionTest("start-unicode-7.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-7.0.0")]
+    public Task start_unicode_7_0_0()
+        => ExecutionTest("start-unicode-7.0.0");
+
+    [Fact(DisplayName = "start-unicode-8.0.0-class-escaped")]
+    public Task start_unicode_8_0_0_class_escaped()
+        => ExecutionTest("start-unicode-8.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-8.0.0-class")]
+    public Task start_unicode_8_0_0_class()
+        => ExecutionTest("start-unicode-8.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-8.0.0-escaped")]
+    public Task start_unicode_8_0_0_escaped()
+        => ExecutionTest("start-unicode-8.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-8.0.0")]
+    public Task start_unicode_8_0_0()
+        => ExecutionTest("start-unicode-8.0.0");
+
+    [Fact(DisplayName = "start-unicode-9.0.0-class-escaped")]
+    public Task start_unicode_9_0_0_class_escaped()
+        => ExecutionTest("start-unicode-9.0.0-class-escaped");
+
+    [Fact(DisplayName = "start-unicode-9.0.0-class")]
+    public Task start_unicode_9_0_0_class()
+        => ExecutionTest("start-unicode-9.0.0-class");
+
+    [Fact(DisplayName = "start-unicode-9.0.0-escaped")]
+    public Task start_unicode_9_0_0_escaped()
+        => ExecutionTest("start-unicode-9.0.0-escaped");
+
+    [Fact(DisplayName = "start-unicode-9.0.0")]
+    public Task start_unicode_9_0_0()
+        => ExecutionTest("start-unicode-9.0.0");
+
+    [Fact(DisplayName = "start-unicode-ltr")]
+    public Task start_unicode_ltr()
+        => ExecutionTest("start-unicode-ltr");
+
+    [Fact(DisplayName = "val-dollar-sign-via-escape-hex")]
+    public Task val_dollar_sign_via_escape_hex()
+        => ExecutionTest("val-dollar-sign-via-escape-hex");
+
+    [Fact(DisplayName = "val-dollar-sign-via-escape-hex4")]
+    public Task val_dollar_sign_via_escape_hex4()
+        => ExecutionTest("val-dollar-sign-via-escape-hex4");
+
+    [Fact(DisplayName = "val-dollar-sign")]
+    public Task val_dollar_sign()
+        => ExecutionTest("val-dollar-sign");
+
+    [Fact(DisplayName = "val-underscore-via-escape-hex")]
+    public Task val_underscore_via_escape_hex()
+        => ExecutionTest("val-underscore-via-escape-hex");
+
+    [Fact(DisplayName = "val-underscore-via-escape-hex4")]
+    public Task val_underscore_via_escape_hex4()
+        => ExecutionTest("val-underscore-via-escape-hex4");
+
+    [Fact(DisplayName = "val-underscore")]
+    public Task val_underscore()
+        => ExecutionTest("val-underscore");
+
+    [Fact(DisplayName = "vals-cjk-escaped")]
+    public Task vals_cjk_escaped()
+        => ExecutionTest("vals-cjk-escaped");
+
+    [Fact(DisplayName = "vals-cjk")]
+    public Task vals_cjk()
+        => ExecutionTest("vals-cjk");
+
+    [Fact(DisplayName = "vals-eng-alpha-lower-via-escape-hex")]
+    public Task vals_eng_alpha_lower_via_escape_hex()
+        => ExecutionTest("vals-eng-alpha-lower-via-escape-hex");
+
+    [Fact(DisplayName = "vals-eng-alpha-lower-via-escape-hex4")]
+    public Task vals_eng_alpha_lower_via_escape_hex4()
+        => ExecutionTest("vals-eng-alpha-lower-via-escape-hex4");
+
+    [Fact(DisplayName = "vals-eng-alpha-upper-via-escape-hex")]
+    public Task vals_eng_alpha_upper_via_escape_hex()
+        => ExecutionTest("vals-eng-alpha-upper-via-escape-hex");
+
+    [Fact(DisplayName = "vals-eng-alpha-upper-via-escape-hex4")]
+    public Task vals_eng_alpha_upper_via_escape_hex4()
+        => ExecutionTest("vals-eng-alpha-upper-via-escape-hex4");
+
+    [Fact(DisplayName = "vals-rus-alpha-lower-via-escape-hex")]
+    public Task vals_rus_alpha_lower_via_escape_hex()
+        => ExecutionTest("vals-rus-alpha-lower-via-escape-hex");
+
+    [Fact(DisplayName = "vals-rus-alpha-lower-via-escape-hex4")]
+    public Task vals_rus_alpha_lower_via_escape_hex4()
+        => ExecutionTest("vals-rus-alpha-lower-via-escape-hex4");
+
+    [Fact(DisplayName = "vals-rus-alpha-lower")]
+    public Task vals_rus_alpha_lower()
+        => ExecutionTest("vals-rus-alpha-lower");
+
+    [Fact(DisplayName = "vals-rus-alpha-upper-via-escape-hex")]
+    public Task vals_rus_alpha_upper_via_escape_hex()
+        => ExecutionTest("vals-rus-alpha-upper-via-escape-hex");
+
+    [Fact(DisplayName = "vals-rus-alpha-upper-via-escape-hex4")]
+    public Task vals_rus_alpha_upper_via_escape_hex4()
+        => ExecutionTest("vals-rus-alpha-upper-via-escape-hex4");
+
+    [Fact(DisplayName = "vals-rus-alpha-upper")]
+    public Task vals_rus_alpha_upper()
+        => ExecutionTest("vals-rus-alpha-upper");
+
+}