+++ tests/Jroc.Test262.Tests/language/statements/variable/dstr/Test262BatchPort20260919Round4ExecutionTests.cs
@@ -1,0 +1,289 @@
+using Jroc.Test262.Tests.language;
+
+namespace Jroc.Test262.Tests.language.statements.variable.dstr;
+
+public class Test262BatchPort20260919Round4ExecutionTests : DiskExecutionTestsBase
+{
+    public Test262BatchPort20260919Round4ExecutionTests() : base("language.statements.variable.dstr") { }
+
+    [Fact(DisplayName = "ary-init-iter-close")]
+    public Task ary_init_iter_close()
+        => ExecutionTest("ary-init-iter-close");
+
+    [Fact(DisplayName = "ary-init-iter-get-err-array-prototype")]
+    public Task ary_init_iter_get_err_array_prototype()
+        => ExecutionTest("ary-init-iter-get-err-array-prototype");
+
+    [Fact(DisplayName = "ary-init-iter-no-close")]
+    public Task ary_init_iter_no_close()
+        => ExecutionTest("ary-init-iter-no-close");
+
+    [Fact(DisplayName = "ary-name-iter-val")]
+    public Task ary_name_iter_val()
+        => ExecutionTest("ary-name-iter-val");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-elem-init")]
+    public Task ary_ptrn_elem_ary_elem_init()
+        => ExecutionTest("ary-ptrn-elem-ary-elem-init");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-elem-iter")]
+    public Task ary_ptrn_elem_ary_elem_iter()
+        => ExecutionTest("ary-ptrn-elem-ary-elem-iter");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-elision-init")]
+    public Task ary_ptrn_elem_ary_elision_init()
+        => ExecutionTest("ary-ptrn-elem-ary-elision-init");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-elision-iter")]
+    public Task ary_ptrn_elem_ary_elision_iter()
+        => ExecutionTest("ary-ptrn-elem-ary-elision-iter");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-empty-init")]
+    public Task ary_ptrn_elem_ary_empty_init()
+        => ExecutionTest("ary-ptrn-elem-ary-empty-init");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-empty-iter")]
+    public Task ary_ptrn_elem_ary_empty_iter()
+        => ExecutionTest("ary-ptrn-elem-ary-empty-iter");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-rest-init")]
+    public Task ary_ptrn_elem_ary_rest_init()
+        => ExecutionTest("ary-ptrn-elem-ary-rest-init");
+
+    [Fact(DisplayName = "ary-ptrn-elem-ary-rest-iter")]
+    public Task ary_ptrn_elem_ary_rest_iter()
+        => ExecutionTest("ary-ptrn-elem-ary-rest-iter");
+
+    [Fact(DisplayName = "ary-ptrn-elem-id-iter-val-err")]
+    public Task ary_ptrn_elem_id_iter_val_err()
+        => ExecutionTest("ary-ptrn-elem-id-iter-val-err");
+
+    [Fact(DisplayName = "ary-ptrn-elem-id-iter-val")]
+    public Task ary_ptrn_elem_id_iter_val()
+        => ExecutionTest("ary-ptrn-elem-id-iter-val");
+
+    [Fact(DisplayName = "ary-ptrn-elem-obj-id-init")]
+    public Task ary_ptrn_elem_obj_id_init()
+        => ExecutionTest("ary-ptrn-elem-obj-id-init");
+
+    [Fact(DisplayName = "ary-ptrn-elem-obj-id")]
+    public Task ary_ptrn_elem_obj_id()
+        => ExecutionTest("ary-ptrn-elem-obj-id");
+
+    [Fact(DisplayName = "ary-ptrn-elem-obj-prop-id-init")]
+    public Task ary_ptrn_elem_obj_prop_id_init()
+        => ExecutionTest("ary-ptrn-elem-obj-prop-id-init");
+
+    [Fact(DisplayName = "ary-ptrn-elem-obj-prop-id")]
+    public Task ary_ptrn_elem_obj_prop_id()
+        => ExecutionTest("ary-ptrn-elem-obj-prop-id");
+
+    [Fact(DisplayName = "ary-ptrn-elem-obj-val-null")]
+    public Task ary_ptrn_elem_obj_val_null()
+        => ExecutionTest("ary-ptrn-elem-obj-val-null");
+
+    [Fact(DisplayName = "ary-ptrn-elem-obj-val-undef")]
+    public Task ary_ptrn_elem_obj_val_undef()
+        => ExecutionTest("ary-ptrn-elem-obj-val-undef");
+
+    [Fact(DisplayName = "ary-ptrn-elision-exhausted")]
+    public Task ary_ptrn_elision_exhausted()
+        => ExecutionTest("ary-ptrn-elision-exhausted");
+
+    [Fact(DisplayName = "ary-ptrn-elision-step-err")]
+    public Task ary_ptrn_elision_step_err()
+        => ExecutionTest("ary-ptrn-elision-step-err");
+
+    [Fact(DisplayName = "ary-ptrn-elision")]
+    public Task ary_ptrn_elision()
+        => ExecutionTest("ary-ptrn-elision");
+
+    [Fact(DisplayName = "ary-ptrn-empty")]
+    public Task ary_ptrn_empty()
+        => ExecutionTest("ary-ptrn-empty");
+
+    [Fact(DisplayName = "ary-ptrn-rest-ary-elem")]
+    public Task ary_ptrn_rest_ary_elem()
+        => ExecutionTest("ary-ptrn-rest-ary-elem");
+
+    [Fact(DisplayName = "ary-ptrn-rest-ary-elision")]
+    public Task ary_ptrn_rest_ary_elision()
+        => ExecutionTest("ary-ptrn-rest-ary-elision");
+
+    [Fact(DisplayName = "ary-ptrn-rest-ary-empty")]
+    public Task ary_ptrn_rest_ary_empty()
+        => ExecutionTest("ary-ptrn-rest-ary-empty");
+
+    [Fact(DisplayName = "ary-ptrn-rest-ary-rest")]
+    public Task ary_ptrn_rest_ary_rest()
+        => ExecutionTest("ary-ptrn-rest-ary-rest");
+
+    [Fact(DisplayName = "ary-ptrn-rest-id-direct")]
+    public Task ary_ptrn_rest_id_direct()
+        => ExecutionTest("ary-ptrn-rest-id-direct");
+
+    [Fact(DisplayName = "ary-ptrn-rest-id-elision-next-err")]
+    public Task ary_ptrn_rest_id_elision_next_err()
+        => ExecutionTest("ary-ptrn-rest-id-elision-next-err");
+
+    [Fact(DisplayName = "ary-ptrn-rest-id-elision")]
+    public Task ary_ptrn_rest_id_elision()
+        => ExecutionTest("ary-ptrn-rest-id-elision");
+
+    [Fact(DisplayName = "ary-ptrn-rest-id-exhausted")]
+    public Task ary_ptrn_rest_id_exhausted()
+        => ExecutionTest("ary-ptrn-rest-id-exhausted");
+
+    [Fact(DisplayName = "ary-ptrn-rest-id-iter-step-err")]
+    public Task ary_ptrn_rest_id_iter_step_err()
+        => ExecutionTest("ary-ptrn-rest-id-iter-step-err");
+
+    [Fact(DisplayName = "ary-ptrn-rest-id-iter-val-err")]
+    public Task ary_ptrn_rest_id_iter_val_err()
+        => ExecutionTest("ary-ptrn-rest-id-iter-val-err");
+
+    [Fact(DisplayName = "ary-ptrn-rest-id")]
+    public Task ary_ptrn_rest_id()
+        => ExecutionTest("ary-ptrn-rest-id");
+
+    [Fact(DisplayName = "ary-ptrn-rest-obj-id")]
+    public Task ary_ptrn_rest_obj_id()
+        => ExecutionTest("ary-ptrn-rest-obj-id");
+
+    [Fact(DisplayName = "ary-ptrn-rest-obj-prop-id")]
+    public Task ary_ptrn_rest_obj_prop_id()
+        => ExecutionTest("ary-ptrn-rest-obj-prop-id");
+
+    [Fact(DisplayName = "obj-init-null")]
+    public Task obj_init_null()
+        => ExecutionTest("obj-init-null");
+
+    [Fact(DisplayName = "obj-init-undefined")]
+    public Task obj_init_undefined()
+        => ExecutionTest("obj-init-undefined");
+
+    [Fact(DisplayName = "obj-ptrn-empty")]
+    public Task obj_ptrn_empty()
+        => ExecutionTest("obj-ptrn-empty");
+
+    [Fact(DisplayName = "obj-ptrn-id-get-value-err")]
+    public Task obj_ptrn_id_get_value_err()
+        => ExecutionTest("obj-ptrn-id-get-value-err");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-fn-name-arrow")]
+    public Task obj_ptrn_id_init_fn_name_arrow()
+        => ExecutionTest("obj-ptrn-id-init-fn-name-arrow");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-fn-name-class")]
+    public Task obj_ptrn_id_init_fn_name_class()
+        => ExecutionTest("obj-ptrn-id-init-fn-name-class");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-fn-name-cover")]
+    public Task obj_ptrn_id_init_fn_name_cover()
+        => ExecutionTest("obj-ptrn-id-init-fn-name-cover");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-fn-name-fn")]
+    public Task obj_ptrn_id_init_fn_name_fn()
+        => ExecutionTest("obj-ptrn-id-init-fn-name-fn");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-fn-name-gen")]
+    public Task obj_ptrn_id_init_fn_name_gen()
+        => ExecutionTest("obj-ptrn-id-init-fn-name-gen");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-skipped")]
+    public Task obj_ptrn_id_init_skipped()
+        => ExecutionTest("obj-ptrn-id-init-skipped");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-throws")]
+    public Task obj_ptrn_id_init_throws()
+        => ExecutionTest("obj-ptrn-id-init-throws");
+
+    [Fact(DisplayName = "obj-ptrn-id-init-unresolvable")]
+    public Task obj_ptrn_id_init_unresolvable()
+        => ExecutionTest("obj-ptrn-id-init-unresolvable");
+
+    [Fact(DisplayName = "obj-ptrn-id-trailing-comma")]
+    public Task obj_ptrn_id_trailing_comma()
+        => ExecutionTest("obj-ptrn-id-trailing-comma");
+
+    [Fact(DisplayName = "obj-ptrn-list-err")]
+    public Task obj_ptrn_list_err()
+        => ExecutionTest("obj-ptrn-list-err");
+
+    [Fact(DisplayName = "obj-ptrn-prop-ary-init")]
+    public Task obj_ptrn_prop_ary_init()
+        => ExecutionTest("obj-ptrn-prop-ary-init");
+
+    [Fact(DisplayName = "obj-ptrn-prop-ary-trailing-comma")]
+    public Task obj_ptrn_prop_ary_trailing_comma()
+        => ExecutionTest("obj-ptrn-prop-ary-trailing-comma");
+
+    [Fact(DisplayName = "obj-ptrn-prop-ary-value-null")]
+    public Task obj_ptrn_prop_ary_value_null()
+        => ExecutionTest("obj-ptrn-prop-ary-value-null");
+
+    [Fact(DisplayName = "obj-ptrn-prop-ary")]
+    public Task obj_ptrn_prop_ary()
+        => ExecutionTest("obj-ptrn-prop-ary");
+
+    [Fact(DisplayName = "obj-ptrn-prop-eval-err")]
+    public Task obj_ptrn_prop_eval_err()
+        => ExecutionTest("obj-ptrn-prop-eval-err");
+
+    [Fact(DisplayName = "obj-ptrn-prop-id-get-value-err")]
+    public Task obj_ptrn_prop_id_get_value_err()
+        => ExecutionTest("obj-ptrn-prop-id-get-value-err");
+
+    [Fact(DisplayName = "obj-ptrn-prop-id-init-skipped")]
+    public Task obj_ptrn_prop_id_init_skipped()
+        => ExecutionTest("obj-ptrn-prop-id-init-skipped");
+
+    [Fact(DisplayName = "obj-ptrn-prop-id-init-throws")]
+    public Task obj_ptrn_prop_id_init_throws()
+        => ExecutionTest("obj-ptrn-prop-id-init-throws");
+
+    [Fact(DisplayName = "obj-ptrn-prop-id-init-unresolvable")]
+    public Task obj_ptrn_prop_id_init_unresolvable()
+        => ExecutionTest("obj-ptrn-prop-id-init-unresolvable");
+
+    [Fact(DisplayName = "obj-ptrn-prop-id-init")]
+    public Task obj_ptrn_prop_id_init()
+        => ExecutionTest("obj-ptrn-prop-id-init");
+
+    [Fact(DisplayName = "obj-ptrn-prop-id-trailing-comma")]
+    public Task obj_ptrn_prop_id_trailing_comma()
+        => ExecutionTest("obj-ptrn-prop-id-trailing-comma");
+
+    [Fact(DisplayName = "obj-ptrn-prop-id")]
+    public Task obj_ptrn_prop_id()
+        => ExecutionTest("obj-ptrn-prop-id");
+
+    [Fact(DisplayName = "obj-ptrn-prop-obj-init")]
+    public Task obj_ptrn_prop_obj_init()
+        => ExecutionTest("obj-ptrn-prop-obj-init");
+
+    [Fact(DisplayName = "obj-ptrn-prop-obj-value-null")]
+    public Task obj_ptrn_prop_obj_value_null()
+        => ExecutionTest("obj-ptrn-prop-obj-value-null");
+
+    [Fact(DisplayName = "obj-ptrn-prop-obj-value-undef")]
+    public Task obj_ptrn_prop_obj_value_undef()
+        => ExecutionTest("obj-ptrn-prop-obj-value-undef");
+
+    [Fact(DisplayName = "obj-ptrn-prop-obj")]
+    public Task obj_ptrn_prop_obj()
+        => ExecutionTest("obj-ptrn-prop-obj");
+
+    [Fact(DisplayName = "obj-ptrn-rest-getter")]
+    public Task obj_ptrn_rest_getter()
+        => ExecutionTest("obj-ptrn-rest-getter");
+
+    [Fact(DisplayName = "obj-ptrn-rest-skip-non-enumerable")]
+    public Task obj_ptrn_rest_skip_non_enumerable()
+        => ExecutionTest("obj-ptrn-rest-skip-non-enumerable");
+
+    [Fact(DisplayName = "obj-ptrn-rest-val-obj")]
+    public Task obj_ptrn_rest_val_obj()
+        => ExecutionTest("obj-ptrn-rest-val-obj");
+
+}